Use CRC instead of PSI OCP4 on GCP instance to run container test

+++ ostree-ng.sh
@@ -26,6 +26,7 @@ INSTALLER_FILENAME=installer.iso
 PROD_REPO_URL=http://192.168.100.1/repo
 PROD_REPO_URL_2="${PROD_REPO_URL}/"
 STAGE_REPO_URL="http://${STAGE_REPO_ADDRESS}:8080/repo/"
+ON_GCP="false"
 
 # Set up temporary files.
 TEMPDIR=$(mktemp -d)
@@ -67,6 +68,10 @@ case "${ID}-${VERSION_ID}" in
         OS_VARIANT="centos-stream8"
         USER_IN_INSTALLER_BP="false"
         ANSIBLE_USER_FOR_BIOS="admin"
+        sudo dnf install -y dmidecode
+        sudo dmidecode -s system-product-name | grep "Google Compute Engine" && ON_GCP="true"
+        # for debugging on openstack
+        # sudo dmidecode -s system-product-name | grep "OpenStack Compute" && ON_GCP="true"
         # ANSIBLE_USER_FOR_BIOS="installeruser"
         ;;
     "centos-9")
@@ -74,12 +79,21 @@ case "${ID}-${VERSION_ID}" in
         OS_VARIANT="centos-stream9"
         USER_IN_INSTALLER_BP="false"
         ANSIBLE_USER_FOR_BIOS="admin"
+        sudo dnf install -y dmidecode
+        sudo dmidecode -s system-product-name | grep "Google Compute Engine" && ON_GCP="true"
+        # for debugging on openstack
+        # sudo dmidecode -s system-product-name | grep "OpenStack Compute" && ON_GCP="true"
         # ANSIBLE_USER_FOR_BIOS="installeruser"
         ;;
     *)
         echo "unsupported distro: ${ID}-${VERSION_ID}"
         exit 1;;
 esac
+
+# Re-define STAGE_OCP4_REPO_URL for CRC on GCP
+if [[ $ON_GCP == "true" ]]; then
+    STAGE_OCP4_REPO_URL="http://${STAGE_OCP4_SERVER_NAME}-${QUAY_REPO_TAG}-default.apps-crc.testing/repo/"
+fi
 
 # Colorful output.
 function greenprint {
@@ -276,6 +290,24 @@ check_result () {
     fi
 }
 
+# Setup CodeReady Containers
+setup_crc () {
+    # Download CRC first
+    wget -4 https://developers.redhat.com/content-gateway/rest/mirror/pub/openshift-v4/clients/crc/latest/crc-linux-amd64.tar.xz
+    sudo tar xf crc-linux-amd64.tar.xz -C /usr/bin --strip-components 1
+    rm -f crc-linux-amd64.tar.xz
+
+    # Configure CRC
+    crc config set consent-telemetry no
+    crc config set disable-update-check true
+    crc config set kubeadmin-password foobar
+    crc config set pull-secret-file "$CRC_PULL_SECRET_FILE"
+
+    # Start CRC
+    crc setup
+    crc start
+}
+
 ###########################################################
 ##
 ## Prepare ostree prod repo and configure stage repo
@@ -350,7 +382,14 @@ sudo rm -f "$IMAGE_FILENAME"
 
 # Run stage repo in OCP4
 greenprint "Running stage repo in OCP4"
-oc login --token="${OCP4_TOKEN}" --server=https://api.ocp-c1.prod.psi.redhat.com:6443 -n rhel-edge --insecure-skip-tls-verify
+if [[ $ON_GCP == "true" ]]; then
+    setup_crc
+    eval $(crc oc-env)
+    oc login https://api.crc.testing:6443 -u kubeadmin -p foobar
+else
+    oc login --token="${OCP4_TOKEN}" --server=https://api.ocp-c1.prod.psi.redhat.com:6443 -n rhel-edge --insecure-skip-tls-verify
+fi
+
 oc process -f tools/edge-stage-server-template.yaml -p EDGE_STAGE_REPO_TAG="${QUAY_REPO_TAG}" -p EDGE_STAGE_SERVER_NAME="${STAGE_OCP4_SERVER_NAME}" | oc apply -f -
 
 for LOOP_COUNTER in $(seq 0 60); do
@@ -368,7 +407,11 @@ sudo ostree --repo="$PROD_REPO" pull --mirror edge-stage-ocp4 "$OSTREE_REF"
 
 # Clean up OCP4
 greenprint "Clean up OCP4"
-oc delete pod,rc,service,route,dc -l app="${STAGE_OCP4_SERVER_NAME}-${QUAY_REPO_TAG}"
+if [[ $ON_GCP == "true" ]]; then
+    crc delete --force
+else
+    oc delete pod,rc,service,route,dc -l app="${STAGE_OCP4_SERVER_NAME}-${QUAY_REPO_TAG}"
+fi
 
 # Clean compose and blueprints.
 greenprint "🧹 Clean up compose"
